if_wg: import latest wg_{noise,cookie}.{c,h}
This seems like quite the daunting diff, however the majority of it is
due to some renamed defines. This fixes a number of bugs that exist in
the old versions and brings these libraries to feature parity with
OpenBSD. It also includes the changes required to call the new library
interfaces.

There are a number of TODO's left in the code that I would like to
resolve before being committed to the main FreeBSD repo. They can all be
fixed by junking the current crypto libraries and doing it properly.

There is a non-zero chance that this has added some subtle bugs due to
assumptions if_wg_session.c has about these libraries, however it
certainly has fixed significatly more bugs than it has created. The next
step is to look over the rest of the code to ensure it is using these
libraries correctly.

Diffs of the updated files and the OpenBSD versions are very readable,
really only due to uma_zone_t vs struct pool and a couple of function
prototypes that are const in FreeBSD and should be const in OpenBSD.