if_wg: replace hacked up crypto non-sense with single-file reference implementation
The crypto port here was incredibly half-baked, and it's very, very hard
to have any assurance about what's happening. It also duplicates tons of
things that should live elsewhere in the kernel. Making if_wg's crypto
work with the rest of the kernel's crypto is a good project for 13.1.
For now, however, the best we can do to make this actually safe to use
is to just nuke the nonsense, and replace it with simplest reference
implementations in a single self-contained file.