if_wg: add ioctl locking
While this removes some of the lockless nature of the ioctls, it is for
the better. This is how we've modeled the locking on other platforms.
This will also mean we can clean up wgc_get and the wg_peer_export
business (if not more).

There will also need to be a bit of work around peer destruction.