if_wg: stop preventing jail teardown
Holding a reference to the initial creds will otherwise prevent the jail
from going away.

Install a handler for PR_METHOD_REMOVE and free the creds of any wg
interface that started in that jail. In the process, we'll mark the
softc as dying so that it can't get up'd in the process.

wg_socket_init() now grabs a hold on the creds under the softc mutex and
manages the lifetime of its local copy to make that a little safer.