ifconfig: if_wg revamp, part 1
Cleanly separate nvl_device and nvl_peer.  Use the latter being
allocated as an indicator that we're in peer mode instead of tracking
that explicitly. Don't bother with extra allocations in peerfinish,
we can just use the address of nvl_peer for our array of one and we're
already allocating a perfectly usable nvl_device.