if_wg: export peers in the order they were added
The peer list is in-fact maintained in reverse order because inserting
at the head of a list is both easier and faster.  Simply reverse it
as we're converting peer_export to nvlists, and note what we're doing.

While we're here, improve the error handling a little bit -- don't
export a partial list of peers, report the no-memory condition.