if_wg: fix socket-outside-vnet behavior
The primary issue here was that we were not setting the correct vnet
for packet delivery -- that of the interface, instead of that of the
socket.

The secondary issue is that the vnet going away could still be stalled
by the wg socket, so start detecting that in our jail removal callback.