if_wg: re-add netgate copyright
Regardless of what code remains and what has been rewritten, it's
simpler to reason about if we just leave it up there. And when you ask
the more basic question, "would this code exist had they not started the
codebase?" the answer is most likely, "no, it wouldn't exist without
them." Besides, the BSD license is already as permissive as it gets, so
it makes no practical difference.