man: make pollution requirements clear for man_check_for_so
The first argument was previously $manpage, but we effectively only
worked with a previously-set $manpage anyways; stop pretending that the
argument matters and just pass the path of interest in.

This previously relied on $path being set in the caller, but that only
worked coincidentally in the search loop for non-paths.  This fixes
.so references when a path to a manpage has been provided rather than
just the name/section of a manpage.

While here, fix the path we pass in while searching for .so references
during the relative/absolute path case.  .so references are w.r.t. to a
mandir, not a mandir/section.

Sponsored by:	Klara, Inc.

(cherry picked from commit 50a321c4a256f4495cab42c14775f9a585593141)