rtld: restore abort for malformed environments while setuid
From 2009 up until bfd4c875a105 ("rtld: avoid use of of getenv(3) ..."),
rtld would abort if a malformed environment was observed in tainted
executions.  The abort was added at the time because we wouldn't be able
to unset any of the sensitive env vars that we try to sanitize.  This
has been rewritten such that that isn't an issue anymore, but a
malformed environment like this can be a good indicator that the caller
is trying to do something hinky; let's restore the previous behavior.

Discussed with:	emaste

(cherry picked from commit a1b186471131f7880e0f1c9758900bc291c79de7)