extres: add support for powerdom
The M1 will have numerous power domains that we need to work with, so
let's add some basic support for this.

OF_device_unregister_xref() has been added so that we can do some
sanity checking on the unregister -- namely, if something else has
clobbered our xref association, we catch that.  We should probably warn
or something there.