pkgbase: update-packages: split out an sign-update-packages target
PKG_ABI is not usable until after WSTAGEDIR has been populated in
stage-packages, therefore we cannot use the value that's picked up in
real-update-packages as it will be out-of-date or the stagedir will not be
populated yet since the `pkg config` invocation is evaluated before any of
this has run.

Split sign-update-packages out to a separate target that we will invoke
after stage-packages. create-packages drops down to a dependency of the
real-update-packages target to be consistent with real-packages, then we
reinvoke in a sub-make so that sign-update-packages can collect any needed
information strictly after it becomes available.

The compromise here is that a missing 'latest' symlink will now be viewed
as a repository needing bootstrapped, but this is not a destructive process;
I consider it harmless.

(cherry picked from commit 6850043be61f4b52683326f2478b1dda6654b5f4)