tools: build: add a rewrite of makeman in lua
The primary benefit of this rewrite is that it parallelizes a number of
the make(1) jobs that it needs to do.  It does so with a very naive
forking model that could likely be improved, but is sufficient for
purposes.  This version also doesn't assume that CWD is sane, and
instead operates relative to the directory the script resides in.

Note that this initial version is only intended to match the output of
the legacy script.

In my horribly performing dev environment, this version runs in 40s
rather than the original ~2 minutes.

On a Mt. Snow machine, this version runs in ~15s rather than the
original ~1m40s.

Differential Revision:	https://reviews.freebsd.org/D39084

(cherry picked from commit 88575dd80581d219f061f3593c8c73341fd448fc)