stand: module: unlink the entire tail when dependencies fail to load
Assume you have loader configured to load modules A, B, C, and D, with
C depending on D and a non-existent E.

When the load command for C fails to find E, it will free both C and D
but only C gets removed from the module list.

Fix it so that we unlink the entire tail of the list.  Anything after
C is, by definition, a dependency on C at this point so we can just
discard it.  If D were loaded before C, it should not move to a position
where it would cause problems here.

(cherry picked from commit 3e73c224cc83a55ea30808d839e5005d2c8ad885)