kern: mountroot: add a mountroot directive to reboot
A future change will add a tunable to control the fallback directive,
currently usually ".ask".  This series is motivated by a desire to avoid
the mountroot prompt and just immediately reboot if we failed to mount
root; nextboot(8) is in use and if we made it far enough to be in a
kernel that's trying to mountroot, odds are greater for a botched
kernel/module than a broken root fs (in my case).

Differential Revision:	https://reviews.freebsd.org/D40314

(cherry picked from commit b7d3c6e8a0e0c41830cf2408091ed3f73824d2ec)