libbe: attempt to remove autocreated mountpoints at unmount time
We use a be_mount.XXXX pattern to mkdtemp(3) when creating these, which
seems reasonably unique enough to just continue using that.  Record the
mountpoint of the root dataset and check the dirname of that for the
auto-creation trait.  There's no sense in this bubbling up an error to
callers, so we'll just ignore an error for now.

Requested by:	manu
Differential Revision:	https://reviews.freebsd.org/D42243

(cherry picked from commit 4b5c263a3571064457f97ee40e12af8347c2dacf)