llvm: implement the bits missing for LSan on FreeBSD
FreeBSD is namely missing a StopTheWorld implementation, but there are other
odds and ends that need to be FreeBSD enabled to work.

The StopTheWorld implementation is generally based on the NetBSD one, which
is in turn based on the Linux one.  We use rfork(RFPROC|RFMEM) to spawn off
a tracer thread that ptraces the target process and analyzes it.

Future work could perhaps use, e.g., a pthread with pthread_suspend_all_np,
but we don't really have a way at the moment to fetch a list of threads or
their registers that we currently obtain with PT_GETREGS.

StopTheWorld implementation and prior work done by Luoqi Chen (luoqi@),
with some portability fixes by kevans@.

Differential Revision:	https://reviews.freebsd.org/D44030

(cherry picked from commit fae22f70d44b560721f85dc7233eb8b66dd70dd5)