kern: net: allow an if_clone_destroyif() without knowing the cloner
Knowing which cloner an interface came from isn't always easy,
especially if a driver has multiple cloner candidates and we need to
find the right cloner for the correct driver in the home vnet.  We have
the information needed, but it's a lot cleaner to just let the clone
infrastructure handle it.

This will be used in a future patch to allow tun(4)/tap(4) devices to
be configured as transient and automatically destroyed on last close of
the device.

Differential Revision:	https://reviews.freebsd.org/D44199

(cherry picked from commit 61a0a9501980855ee66aa6fb76f45fb2fd05748d)