kern: squashfs: simplify decompressor support a bit
Just use a single array of decompression definitions instead of having to
define each one multiple times based on whether it's supported or not with
the current kernel configuration.  Nix the "unknown" entry, just return NULL
so that the caller can detect and use a slightly different message if the
type's not recognized.