caroot: note the source license we extracted from
NSS and certdata.txt, which we extract the root store from, are licensed
under the MPL2.  It's not clear that certificate data like this may be
copyrighted in the first place, but in the interest of being a good
consumer of NSS data note the source license in extracted certs.

Differential Revision:	https://reviews.freebsd.org/D45067

(cherry picked from commit f223b94f13534956f6a0340b9f191def9f1ef5a3)